Move functions related to s3 urls to a new pkg
